Fix all syntax errors for unit tests

+++ app/src/test/java/com/gmail/bogumilmecel2/fitnessappv2/feature_diary/domain/use_cases/CalculateSkipUseCaseTest.kt
@@ -19,7 +19,7 @@ class CalculateSkipUseCaseTest {
 
     private fun testMethod(
         page: Int,
-        expectedValue: Int
+        expectedValue: Long
     ) {
         assertEquals(
             expected = expectedValue,
